+++ roseau/load_flow/models/transformers/transformers.py
@@ -10,7 +10,7 @@ from roseau.load_flow.models.buses import Bus
 from roseau.load_flow.models.transformers.parameters import TransformerParameters
 from roseau.load_flow.typing import Id, JsonDict
 from roseau.load_flow.units import Q_, ureg_wraps
-from roseau.load_flow.utils import find_stack_level
+from roseau.load_flow.utils import deprecate_renamed_parameters, find_stack_level
 from roseau.load_flow_engine.cy_engine import (
     CyCenterTransformer,
     CySingleTransformer,
@@ -32,23 +32,28 @@ class Transformer(AbstractBranch):
 
     - P-P-P or P-P-P-N: ``"abc"``, ``"abcn"`` (three-phase transformer)
     - P-P or P-N: ``"ab"``, ``"bc"``, ``"ca"``, ``"an"``, ``"bn"``, ``"cn"`` (single-phase
-      transformer or primary of center-tapped transformer)
-    - P-P-N: ``"abn"``, ``"bcn"``, ``"can"`` (secondary of center-tapped transformer)
+      transformer or HV side of center-tapped transformer)
+    - P-P-N: ``"abn"``, ``"bcn"``, ``"can"`` (LV side of center-tapped transformer)
     """
     _allowed_phases_three = frozenset({"abc", "abcn"})
     _allowed_phases_single = frozenset({"ab", "bc", "ca", "an", "bn", "cn"})
-    _allowed_phases_center_secondary = frozenset({"abn", "bcn", "can"})
+    _allowed_phases_center_lv_side = frozenset({"abn", "bcn", "can"})
 
+    @deprecate_renamed_parameters(
+        {"bus1": "bus_hv", "bus2": "bus_lv", "phases1": "phases_hv", "phases2": "phases_lv"},
+        version="0.12.0",
+        category=DeprecationWarning,
+    )
     def __init__(
         self,
         id: Id,
-        bus1: Bus,
-        bus2: Bus,
+        bus_hv: Bus,
+        bus_lv: Bus,
         *,
         parameters: TransformerParameters,
         tap: float = 1.0,
-        phases1: str | None = None,
-        phases2: str | None = None,
+        phases_hv: str | None = None,
+        phases_lv: str | None = None,
         max_loading: float | Q_[float] = 1,
         geometry: BaseGeometry | None = None,
     ) -> None:
@@ -56,13 +61,13 @@ class Transformer(AbstractBranch):
 
         Args:
             id:
-                A unique ID of the transformer in the network branches.
+                A unique ID of the transformer in the network transformers.
 
-            bus1:
-                Bus to connect the first extremity of the transformer.
+            bus_hv:
+                Bus to connect the HV side of the transformer.
 
-            bus2:
-                Bus to connect the first extremity of the transformer.
+            bus_lv:
+                Bus to connect the LV side of the transformer.
 
             tap:
                 The tap of the transformer, for example 1.02.
@@ -71,14 +76,14 @@ class Transformer(AbstractBranch):
                 Parameters defining the electrical model of the transformer. This is an instance of
                 the :class:`TransformerParameters` class and can be used by multiple transformers.
 
-            phases1:
-                The phases of the first extremity of the transformer. A string like ``"abc"`` or
-                ``"abcn"`` etc. The order of the phases is important. For a full list of supported
-                phases, see the class attribute :attr:`allowed_phases`. All phases must be present
-                in the connected bus. By default, determined from the transformer type.
+            phases_hv:
+                The phases of the HV side of the transformer. A string like ``"abc"`` or ``"abcn"``
+                etc. The order of the phases is important. For a full list of supported phases, see
+                the class attribute :attr:`allowed_phases`. All phases must be present in the
+                connected bus. By default, determined from the transformer type.
 
-            phases2:
-                The phases of the second extremity of the transformer. See ``phases1``.
+            phases_lv:
+                The phases of the LV side of the transformer. Similar to ``phases_hv``.
 
             max_loading:
                 The maximum loading of the transformer (unitless). It is used with the `sn` of the
@@ -90,25 +95,25 @@ class Transformer(AbstractBranch):
                 The geometry of the transformer.
         """
         if parameters.type == "single-phase":
-            phases1, phases2 = self._compute_phases_single(
-                id=id, bus1=bus1, bus2=bus2, phases1=phases1, phases2=phases2
+            phases_hv, phases_lv = self._compute_phases_single(
+                id=id, bus_hv=bus_hv, bus_lv=bus_lv, phases_hv=phases_hv, phases_lv=phases_lv
             )
         elif parameters.type == "center-tapped":
-            phases1, phases2 = self._compute_phases_center(
-                id=id, bus1=bus1, bus2=bus2, phases1=phases1, phases2=phases2
+            phases_hv, phases_lv = self._compute_phases_center(
+                id=id, bus_hv=bus_hv, bus_lv=bus_lv, phases_hv=phases_hv, phases_lv=phases_lv
             )
         else:
-            phases1, phases2 = self._compute_phases_three(
-                id=id, bus1=bus1, bus2=bus2, parameters=parameters, phases1=phases1, phases2=phases2
+            phases_hv, phases_lv = self._compute_phases_three(
+                id=id, bus_hv=bus_hv, bus_lv=bus_lv, parameters=parameters, phases_hv=phases_hv, phases_lv=phases_lv
             )
 
-        super().__init__(id=id, bus1=bus1, bus2=bus2, phases1=phases1, phases2=phases2, geometry=geometry)
+        super().__init__(id=id, bus1=bus_hv, bus2=bus_lv, phases1=phases_hv, phases2=phases_lv, geometry=geometry)
         self.tap = tap
         self._parameters = parameters
         self.max_loading = max_loading
 
         z2, ym, k = parameters._z2, parameters._ym, parameters._k
-        clock, orientation = parameters._phase_displacement, parameters.orientation
+        clock, orientation = parameters.clock, parameters.orientation
         self._cy_element: CyTransformer
         if parameters.type == "single-phase":
             self._cy_element = CySingleTransformer(z2=z2, ym=ym, k=k * orientation * tap)
@@ -118,8 +123,8 @@ class Transformer(AbstractBranch):
             self._cy_element = CyThreePhaseTransformer(
                 n1=parameters._n1,
                 n2=parameters._n2,
-                prim=parameters.winding1[0],
-                sec=parameters.winding2[0],
+                whv=parameters.whv[0],
+                wlv=parameters.wlv[0],
                 z2=z2,
                 ym=ym,
                 k=k * tap,
@@ -129,10 +134,30 @@ class Transformer(AbstractBranch):
 
     def __repr__(self) -> str:
         return (
-            f"<{type(self).__name__}: id={self.id!r}, bus1={self.bus1.id!r}, bus2={self.bus2.id!r}, "
-            f"phases1={self.phases1!r}, phases2={self.phases2!r}, tap={self.tap:f}, "
+            f"<{type(self).__name__}: id={self.id!r}, bus_hv={self.bus_hv.id!r}, bus_lv={self.bus_lv.id!r}, "
+            f"phases_hv={self.phases_hv!r}, phases_lv={self.phases_lv!r}, tap={self.tap:f}, "
             f"max_loading={self._max_loading:f}>"
         )
+
+    @property
+    def bus_hv(self) -> Bus:
+        """The bus on the high voltage side of the transformer."""
+        return self._bus1
+
+    @property
+    def bus_lv(self) -> Bus:
+        """The bus on the low voltage side of the transformer."""
+        return self._bus2
+
+    @property
+    def phases_hv(self) -> str:
+        """The phases of the high voltage side of the transformer."""
+        return self._phases1
+
+    @property
+    def phases_lv(self) -> str:
+        """The phases of the low voltage side of the transformer."""
+        return self._phases2
 
     @property
     def tap(self) -> float:
@@ -206,29 +231,29 @@ class Transformer(AbstractBranch):
     def _compute_phases_three(
         self,
         id: Id,
-        bus1: Bus,
-        bus2: Bus,
+        bus_hv: Bus,
+        bus_lv: Bus,
         parameters: TransformerParameters,
-        phases1: str | None,
-        phases2: str | None,
+        phases_hv: str | None,
+        phases_lv: str | None,
     ) -> tuple[str, str]:
-        w1 = parameters.winding1
-        w2 = parameters.winding2
-        clock = parameters.phase_displacement
+        whv = parameters.whv
+        wlv = parameters.wlv
+        clock = parameters.clock
 
-        w1_has_neutral = w1.endswith("N")
-        if phases1 is None:
-            phases1 = "abcn" if w1_has_neutral else "abc"
-            phases1 = "".join(p for p in bus1.phases if p in phases1)
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases1=phases1)
+        w1_has_neutral = whv.endswith("N")
+        if phases_hv is None:
+            phases_hv = "abcn" if w1_has_neutral else "abc"
+            phases_hv = "".join(p for p in bus_hv.phases if p in phases_hv)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases_hv=phases_hv)
         else:
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases1=phases1)
-            self._check_bus_phases(id=id, bus=bus1, phases1=phases1)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases_hv=phases_hv)
+            self._check_bus_phases(id=id, bus=bus_hv, phases_hv=phases_hv)
             transformer_phases = "abcn" if w1_has_neutral else "abc"
-            phases_not_in_transformer = set(phases1) - set(transformer_phases)
+            phases_not_in_transformer = set(phases_hv) - set(transformer_phases)
             if phases_not_in_transformer:
-                if phases_not_in_transformer == {"n"} and w1.startswith(("Y", "Z")):
-                    correct_vg = f"{w1}N{w2}{clock}"
+                if phases_not_in_transformer == {"n"} and whv.startswith(("Y", "Z")):
+                    correct_vg = f"{whv}N{wlv}{clock}"
                     warnings.warn(
                         f"Transformer {id!r} with vector group '{parameters.vg}' does not have a "
                         f"brought out neutral on the HV side. The neutral phase 'n' is ignored. If "
@@ -237,25 +262,25 @@ class Transformer(AbstractBranch):
                         FutureWarning,
                         stacklevel=find_stack_level(),
                     )
-                    phases1 = phases1.replace("n", "")
+                    phases_hv = phases_hv.replace("n", "")
                 else:
-                    msg = f"Phases (1) {phases1!r} of transformer {id!r} are not compatible with its winding {w1!r}."
+                    msg = f"HV phases {phases_hv!r} of transformer {id!r} are not compatible with its winding {whv!r}."
                     logger.error(msg)
                     raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
 
-        w2_has_neutral = w2.endswith("n")
-        if phases2 is None:
-            phases2 = "abcn" if w2_has_neutral else "abc"
-            phases2 = "".join(p for p in bus2.phases if p in phases2)
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases2=phases2)
+        w2_has_neutral = wlv.endswith("n")
+        if phases_lv is None:
+            phases_lv = "abcn" if w2_has_neutral else "abc"
+            phases_lv = "".join(p for p in bus_lv.phases if p in phases_lv)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases_lv=phases_lv)
         else:
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases2=phases2)
-            self._check_bus_phases(id=id, bus=bus2, phases2=phases2)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_three, phases_lv=phases_lv)
+            self._check_bus_phases(id=id, bus=bus_lv, phases_lv=phases_lv)
             transformer_phases = "abcn" if w2_has_neutral else "abc"
-            phases_not_in_transformer = set(phases2) - set(transformer_phases)
+            phases_not_in_transformer = set(phases_lv) - set(transformer_phases)
             if phases_not_in_transformer:
-                if phases_not_in_transformer == {"n"} and w2.startswith(("y", "z")):
-                    correct_vg = f"{w1}{w2}n{clock}"
+                if phases_not_in_transformer == {"n"} and wlv.startswith(("y", "z")):
+                    correct_vg = f"{whv}{wlv}n{clock}"
                     warnings.warn(
                         f"Transformer {id!r} with vector group '{parameters.vg}' does not have a "
                         f"brought out neutral on the LV side. The neutral phase 'n' is ignored. If "
@@ -264,77 +289,84 @@ class Transformer(AbstractBranch):
                         FutureWarning,
                         stacklevel=find_stack_level(),
                     )
-                    phases2 = phases2.replace("n", "")
+                    phases_lv = phases_lv.replace("n", "")
                 else:
-                    msg = f"Phases (2) {phases2!r} of transformer {id!r} are not compatible with its winding {w2!r}."
+                    msg = f"LV phases {phases_lv!r} of transformer {id!r} are not compatible with its winding {wlv!r}."
                     logger.error(msg)
                     raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
 
-        return phases1, phases2
+        return phases_hv, phases_lv
 
     def _compute_phases_single(
-        self, id: Id, bus1: Bus, bus2: Bus, phases1: str | None, phases2: str | None
+        self, id: Id, bus_hv: Bus, bus_lv: Bus, phases_hv: str | None, phases_lv: str | None
     ) -> tuple[str, str]:
-        if phases1 is None:
-            phases1 = "".join(p for p in bus1.phases if p in bus2.phases)  # can't use set because order is important
-            phases1 = phases1.replace("ac", "ca")
-            if phases1 not in self._allowed_phases_single:
-                msg = f"Phases (1) of transformer {id!r} cannot be deduced from the buses, they need to be specified."
+        if phases_hv is None:
+            phases_hv = "".join(
+                p for p in bus_hv.phases if p in bus_lv.phases
+            )  # can't use set because order is important
+            phases_hv = phases_hv.replace("ac", "ca")
+            if phases_hv not in self._allowed_phases_single:
+                msg = f"HV phases of transformer {id!r} cannot be deduced from the buses, they need to be specified."
                 logger.error(msg)
                 raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
         else:
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_single, phases1=phases1)
-            self._check_bus_phases(id=id, bus=bus1, phases1=phases1)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_single, phases_hv=phases_hv)
+            self._check_bus_phases(id=id, bus=bus_hv, phases_hv=phases_hv)
 
-        if phases2 is None:
-            phases2 = "".join(p for p in bus1.phases if p in bus2.phases)  # can't use set because order is important
-            phases2 = phases2.replace("ac", "ca")
-            if phases2 not in self._allowed_phases_single:
-                msg = f"Phases (2) of transformer {id!r} cannot be deduced from the buses, they need to be specified."
+        if phases_lv is None:
+            phases_lv = "".join(
+                p for p in bus_hv.phases if p in bus_lv.phases
+            )  # can't use set because order is important
+            phases_lv = phases_lv.replace("ac", "ca")
+            if phases_lv not in self._allowed_phases_single:
+                msg = f"LV phases of transformer {id!r} cannot be deduced from the buses, they need to be specified."
                 logger.error(msg)
                 raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
         else:
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_single, phases2=phases2)
-            self._check_bus_phases(id=id, bus=bus2, phases2=phases2)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_single, phases_lv=phases_lv)
+            self._check_bus_phases(id=id, bus=bus_lv, phases_lv=phases_lv)
 
-        return phases1, phases2
+        return phases_hv, phases_lv
 
     def _compute_phases_center(
-        self, id: Id, bus1: Bus, bus2: Bus, phases1: str | None, phases2: str | None
+        self, id: Id, bus_hv: Bus, bus_lv: Bus, phases_hv: str | None, phases_lv: str | None
     ) -> tuple[str, str]:
-        if phases1 is None:
-            phases1 = "".join(p for p in bus2.phases if p in bus1.phases and p != "n")
-            phases1 = phases1.replace("ac", "ca")
-            if phases1 not in self._allowed_phases_single:
-                msg = f"Phases (1) of transformer {id!r} cannot be deduced from the buses, they need to be specified."
+        if phases_hv is None:
+            phases_hv = "".join(p for p in bus_lv.phases if p in bus_hv.phases and p != "n")
+            phases_hv = phases_hv.replace("ac", "ca")
+            if phases_hv not in self._allowed_phases_single:
+                msg = f"HV phases of transformer {id!r} cannot be deduced from the buses, they need to be specified."
                 logger.error(msg)
                 raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
         else:
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_single, phases1=phases1)
-            self._check_bus_phases(id=id, bus=bus1, phases1=phases1)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_single, phases_hv=phases_hv)
+            self._check_bus_phases(id=id, bus=bus_hv, phases_hv=phases_hv)
 
-        if phases2 is None:
-            phases2 = "".join(p for p in bus2.phases if p in bus1.phases or p == "n")
-            if phases2 not in self._allowed_phases_center_secondary:
-                msg = f"Phases (2) of transformer {id!r} cannot be deduced from the buses, they need to be specified."
+        if phases_lv is None:
+            phases_lv = "".join(p for p in bus_lv.phases if p in bus_hv.phases or p == "n")
+            if phases_lv not in self._allowed_phases_center_lv_side:
+                msg = f"LV phases of transformer {id!r} cannot be deduced from the buses, they need to be specified."
                 logger.error(msg)
                 raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
         else:
-            self._check_phases(id=id, allowed_phases=self._allowed_phases_center_secondary, phases2=phases2)
-            self._check_bus_phases(id=id, bus=bus2, phases2=phases2)
+            self._check_phases(id=id, allowed_phases=self._allowed_phases_center_lv_side, phases_lv=phases_lv)
+            self._check_bus_phases(id=id, bus=bus_lv, phases_lv=phases_lv)
 
-        return phases1, phases2
+        return phases_hv, phases_lv
 
     @staticmethod
     def _check_bus_phases(id: Id, bus: Bus, **kwargs: str) -> None:
-        name, phases = kwargs.popitem()  # phases1 or phases2
-        name = "Phases (1)" if name == "phases1" else "Phases (2)"
+        name, phases = kwargs.popitem()  # phases_hv or phases_lv
+        side = "HV" if name == "phases_hv" else "LV"
         phases_not_in_bus = set(phases) - set(bus.phases)
         if phases_not_in_bus:
-            msg = (
-                f"{name} {sorted(phases_not_in_bus)} of transformer {id!r} are not in phases "
-                f"{bus.phases!r} of bus {bus.id!r}."
-            )
+            if len(phases_not_in_bus) == 1:
+                ph = f"phase {next(iter(phases_not_in_bus))!r}"
+                be = "is"
+            else:
+                ph = f"phases {sorted(phases_not_in_bus)}"
+                be = "are"
+            msg = f"{side} {ph} of transformer {id!r} {be} not in phases {bus.phases!r} of its {side} bus {bus.id!r}."
             logger.error(msg)
             raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
 
@@ -358,7 +390,7 @@ class Transformer(AbstractBranch):
     @property
     def res_violated(self) -> bool:
         """Whether the transformer power loading exceeds its maximal loading."""
-        # True if either the primary or secondary is overloaded
+        # True if either the HV or LV side is overloaded
         loading = self._res_loading_getter(warning=True)
         return bool(loading > self._max_loading)
 
@@ -377,8 +409,8 @@ class Transformer(AbstractBranch):
         currents1, currents2 = self._res_currents_getter(warning)
         results = {
             "id": self.id,
-            "phases1": self.phases1,
-            "phases2": self.phases2,
+            "phases1": self.phases_hv,
+            "phases2": self.phases_lv,
             "currents1": [[i.real, i.imag] for i in currents1],
             "currents2": [[i.real, i.imag] for i in currents2],
         }